decouple write utterances file
